requirements.summariser.txt
anthropic==0.49.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
requests==2.32.3
